# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.DistritoReniec:
    properties:
      idDistrito:
        type: integer
      idProvincia:
        type: integer
      idReniec:
        type: integer
      nombre:
        type: string
    type: object
...
  domain.ListadoUpsItem:
    properties:
      codUps:
        type: string
      descripcion:
        type: string
    type: object
  domain.ListadoUpsResponse:
    properties:
      codigo:
        type: string
      datos:
        items:
          $ref: '#/definitions/domain.ListadoUpsItem'
        type: array
      mensaje:
        type: string
    type: object
...
  domain.ReferenciaPorMes:
    properties:
      cantidad:
        type: integer
      estado:
        type: string
      mes:
        type: integer
      nombreMes:
        type: string
    type: object
...
  domain.Ups:
    properties:
      codigo:
        type: string
      descripcion:
        type: string
    type: object
...
  refconhttp.apiError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  refconhttp.apiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/refconhttp.apiError'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal,
    persiste en SQL Server 2022.
  title: Galenos Pro Appointments API
  version: "1.0"
paths:
...
  /dashrefcon/referencias:
    get:
      consumes:
      - application/json
      description: Retorna la cantidad de referencias/contrarreferencias agrupadas
        por mes y estado para el dashboard de barras
      parameters:
      - description: Mes (0 = todos los meses)
        in: query
        name: mes
        type: integer
      - description: Año de la referencia
        in: query
        name: anio
        required: true
        type: integer
      - description: Estado de la referencia (3 ACEPTADO, 4 RECHAZADO, 5 PACIENTE
          RECIBIDO, 7 PACIENTE CITADO, 8 CONTRAREFERIDO)
        in: query
        name: estado
        type: integer
      - description: Código de la UPS
        in: query
        name: ups
        type: string
      - description: Opción del reporte (1 total, 2 por estado, 3 por estado filtrado,
          4 por estado y UPS)
        in: query
        name: opcion
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/refconhttp.apiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ReferenciaPorMes'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Listar referencias por mes y año
      tags:
      - DashRefCon
  /dashrefcon/ups:
    get:
      description: Retorna el catálogo de Unidades Productoras de Servicios (UPS)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/refconhttp.apiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Ups'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Listar UPS
      tags:
      - DashRefCon
  /dashrefcon/upss/{codigoRenipress}:
    get:
      description: Consulta el listado de Unidades Productoras de Servicios (UPS)
        del establecimiento indicado por su código RENIPRESS en el servicio REST del
        MINSA
      parameters:
      - description: Código RENIPRESS del establecimiento
        in: path
        name: codigoRenipress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/refconhttp.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ListadoUpsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Listar UPS de un establecimiento en MINSA
      tags:
      - DashRefCon
...
  /refcon/distritos-reniec/{idReniec}:
    get:
      description: Retorna el distrito asociado a un código RENIEC (tabla Distritos)
      parameters:
      - description: IdReniec del distrito
        in: path
        name: idReniec
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/refconhttp.apiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.DistritoReniec'
                  type: array
              type: object
      summary: Listar distritos por IdReniec
      tags:
      - RefCon
...
swagger: "2.0"